+++ src/main/java/ru/made/ex/one/IrSwapTrade.java
@@ -2,7 +2,7 @@ package ru.made.ex.one;
 
 public class IrSwapTrade extends BaseTrade {
 
-    public IrSwapTrade(Float price) {
+    public IrSwapTrade(double price) {
         super(price);
     }
 }
